add different profile fragment classes (#126)
* add different profile fragment classes

* invisible add media button

* delete testingactivity

* add tests

* fix

* add more timeout

* tests pass locally?

* squish layout to make it pass on the emulator

* ahhh

* more squish

* fix test
