+++ app/src/main/AndroidManifest.xml
@@ -22,9 +22,6 @@
             android:name=".NewProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".TestingActivity"
-            android:exported="false" />
-        <activity
             android:name=".GreetingActivity"
             android:exported="false" />
         <activity
