+++ app/src/main/res/layout/layout_collection_list.xml
@@ -6,6 +6,7 @@
 
   <Button
     android:id="@+id/add_collection_button"
+    android:visibility="gone"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:layout_marginTop="10dp"
