+++ app/src/main/res/values/strings.xml
@@ -32,6 +32,8 @@
   <string name="cancel">Cancel</string>
   <string name="film">Film</string>
   <string name="books">Books</string>
+  <string name="followers_button">%d Followers</string>
+  <string name="following_button">%d Following</string>
 
   <!-- Search Categories -->
   <string name="search_category_movies">Movies</string>
